#7823: Add Falcon7b to device perf pipeline
