+++ models/demos/falcon7b/tests/test_perf_falcon.py
@@ -2,371 +2,21 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import torch
 import pytest
-from loguru import logger
-import numpy as np
-from sklearn.metrics import top_k_accuracy_score
 
-import tt_lib
-from models.demos.falcon7b.reference.hf_modeling_falcon import (
-    FalconForCausalLM,
-)
-from models.demos.falcon7b.tt.falcon_causallm import TtFalconCausalLM
-
-# TODO: Remove this?
-from models.demos.falcon7b.tt.falcon_common import (
-    PytorchFalconCausalLM,
-)
-
+from models.demos.falcon7b.tests.run_falcon_end_to_end import run_test_FalconCausalLM_end_to_end
 from models.demos.falcon7b.tt.model_config import (
     get_model_config,
 )
-from models.demos.falcon7b.tests.test_utils import (
-    get_rand_falcon_inputs,
-    concat_device_out_layer_present,
-    load_hf_model,
-)
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import (
-    get_atol_rtol_pcc,
-)
 
 from models.utility_functions import (
-    tt_tensors_to_torch_tensors,
-    profiler,
-    enable_persistent_kernel_cache,
-    disable_persistent_kernel_cache,
     disable_compilation_reports,
+    disable_persistent_kernel_cache,
     is_e75,
-    is_wormhole_b0,
     skip_for_grayskull,
     skip_for_wormhole_b0,
     get_devices_for_t3000,
 )
-from models.perf.perf_utils import prep_perf_report
-
-
-def get_inputs_on_device(llm_mode, tt_FalconCausalLM, model_input, kv_cache_len, seq_len, batch, kv_len):
-    if llm_mode == "prefill":
-        tt_input_ids, tt_attention_mask = zip(
-            *[
-                tt_FalconCausalLM.model_preprocessing(
-                    llm_mode, model_input[i::batch], kv_cache_len, num_input_tokens=seq_len
-                )
-                for i in range(batch)
-            ]
-        )
-    elif llm_mode == "decode":
-        tt_input_ids, tt_attention_mask = tt_FalconCausalLM.model_preprocessing(
-            llm_mode, model_input, kv_cache_len, num_input_tokens=kv_len
-        )
-    return tt_input_ids, tt_attention_mask
-
-
-# TODO: Replace this with actual Falcon application-level tests
-def run_test_FalconCausalLM_end_to_end(
-    devices,
-    model_version,
-    llm_mode,
-    batch,
-    seq_len,
-    kv_cache_len,
-    num_layers,
-    expected_pccs,
-    model_config,
-    model_config_str,
-    tt_cache_path,
-    model_location_generator,
-    expected_inference_time,
-    async_mode=False,
-):
-    # Clear global profiler state before starting measurements
-    profiler.clear()
-
-    num_devices = len(devices)
-    global_batch = batch * num_devices
-
-    profiler.start("hugging_face_model_setup")
-    hugging_face_reference_model, state_dict = load_hf_model(model_location_generator, model_version)
-    configuration = hugging_face_reference_model.config
-    pytorch_FalconCausalLM = PytorchFalconCausalLM(hugging_face_reference_model, num_layers)
-    profiler.end("hugging_face_model_setup")
-
-    # Prepare input ------------------------------------------------------------------------
-    torch.manual_seed(0)
-    base_url = ""
-    max_position_embeddings = 2048
-    head_dim = configuration.hidden_size // configuration.num_attention_heads
-    use_cache = True
-
-    if True:
-        model_input = torch.arange(seq_len * global_batch).reshape(global_batch, seq_len)
-    else:
-        # batch identical sequences for debugging
-        model_input = torch.stack([torch.arange(seq_len)] * global_batch).reshape(global_batch, seq_len)
-
-    # Generate dummy kv_cache --------------------------------------------------------------
-    (
-        past_key_values,
-        tt_layer_past,
-        kv_len,
-    ) = get_rand_falcon_inputs(
-        llm_mode,
-        seq_len,
-        batch,
-        kv_cache_len,
-        devices,
-        global_batch,
-        head_dim,
-        max_position_embeddings,
-        configuration,
-        model_config,
-        num_layers=num_layers,
-        generate_attention_inputs=False,
-    )
-
-    profiler.start("TtFalcon_model_setup")
-    tt_FalconCausalLM = TtFalconCausalLM(
-        devices,
-        state_dict,
-        base_url,
-        num_layers,
-        configuration,
-        max_position_embeddings,
-        model_config,
-        tt_cache_path,
-        seq_len,
-    )
-    profiler.end("TtFalcon_model_setup")
-
-    profiler.start("processing_of_input")
-    # TODO: Generate attention_mask on device
-    tt_input_ids, tt_attention_mask = get_inputs_on_device(
-        llm_mode, tt_FalconCausalLM, model_input, kv_cache_len, seq_len, batch, kv_len
-    )
-    profiler.end("processing_of_input")
-
-    # First run to fill compile cache ----------------------------------------------------
-    logger.info(f"Running Falcon model once to fill caches -> disable profiler")
-    profiler.disable()
-
-    # Use force enable to only record this profiler call while others are disabled
-    profiler.start("first_model_run_with_compile", force_enable=True)
-    if llm_mode == "prefill":
-        tt_outs = []
-        # Device transfer time is included in model run time for prefill
-        tt_input_ids, tt_attention_mask = get_inputs_on_device(
-            llm_mode, tt_FalconCausalLM, model_input, kv_cache_len, seq_len, batch, kv_len
-        )
-        for user_id in range(batch):
-            tt_out, tt_layer_present = tt_FalconCausalLM(
-                input_ids=tt_input_ids[user_id],
-                llm_mode=llm_mode,
-                attention_mask=tt_attention_mask[user_id],
-                user_id=user_id,
-                layer_past=tt_layer_past,
-                layer_past_len=kv_cache_len,
-                use_cache=use_cache,
-            )
-            tt_outs.append(tt_out)
-        tt_out = tt_outs
-
-    elif llm_mode == "decode":
-        tt_out, tt_layer_present = tt_FalconCausalLM(
-            input_ids=tt_input_ids,
-            llm_mode=llm_mode,
-            attention_mask=tt_attention_mask,
-            layer_past=tt_layer_past,
-            layer_past_len=kv_cache_len,
-            use_cache=use_cache,
-        )
-    for device in devices:
-        tt_lib.device.Synchronize(device)
-    profiler.end("first_model_run_with_compile", force_enable=True)
-
-    # Dump device profiler data before second run to avoid exceeding profiler memory limits when using tracy
-    for device in devices:
-        tt_lib.device.DumpDeviceProfiler(device)
-
-    del tt_out
-    del tt_layer_past
-    del tt_layer_present
-    del tt_input_ids
-    del tt_attention_mask
-
-    # Re-generate dummy kv_cache ------------------------------------------------------------
-    (
-        past_key_values,
-        tt_layer_past,
-        kv_len,
-    ) = get_rand_falcon_inputs(
-        llm_mode,
-        seq_len,
-        batch,
-        kv_cache_len,
-        devices,
-        global_batch,
-        head_dim,
-        max_position_embeddings,
-        configuration,
-        model_config,
-        num_layers=num_layers,
-        generate_attention_inputs=False,
-    )
-
-    # Prepare reference output -----------------------------------------------------------
-
-    profiler.start("hugging_face_reference_model")
-    pytorch_out, pytorch_layer_present = pytorch_FalconCausalLM(
-        input_ids=model_input, past_key_values=past_key_values, use_cache=use_cache
-    )
-    profiler.end("hugging_face_reference_model")
-
-    # Second run for perf ----------------------------------------------------------------
-
-    logger.info(f"Enable profiler and enable binary and compile cache")
-    profiler.enable()
-    enable_persistent_kernel_cache()
-
-    # Regenerate input ids and attention_mask on device
-    tt_input_ids, tt_attention_mask = get_inputs_on_device(
-        llm_mode, tt_FalconCausalLM, model_input, kv_cache_len, seq_len, batch, kv_len
-    )
-
-    profiler.start(f"model_run_for_inference")
-    if llm_mode == "prefill":
-        tt_outs = []
-        # Device transfer time is included in model run time for prefill
-        tt_input_ids, tt_attention_mask = get_inputs_on_device(
-            llm_mode, tt_FalconCausalLM, model_input, kv_cache_len, seq_len, batch, kv_len
-        )
-        for user_id in range(batch):
-            tt_out, tt_layer_present = tt_FalconCausalLM(
-                input_ids=tt_input_ids[user_id],
-                llm_mode=llm_mode,
-                attention_mask=tt_attention_mask[user_id],
-                user_id=user_id,
-                layer_past=tt_layer_past,
-                layer_past_len=kv_cache_len,
-                use_cache=use_cache,
-            )
-            tt_outs.append(tt_out)
-
-    elif llm_mode == "decode":
-        tt_out, tt_layer_present = tt_FalconCausalLM(
-            input_ids=tt_input_ids,
-            llm_mode=llm_mode,
-            attention_mask=tt_attention_mask,
-            layer_past=tt_layer_past,
-            layer_past_len=kv_cache_len,
-            use_cache=use_cache,
-        )
-    for device in devices:
-        tt_lib.device.Synchronize(device)
-    profiler.end(f"model_run_for_inference")
-
-    if llm_mode == "prefill":
-        tt_out_tmp = torch.zeros(global_batch, seq_len, configuration.vocab_size)  # Output tensor to overwrite
-        for user_id, tt_out in enumerate(tt_outs):
-            # Get outputs from all devices
-            tt_out_tmp[user_id::batch] = torch.concat(
-                [tt_out_torch.squeeze(1) for tt_out_torch in tt_tensors_to_torch_tensors(tt_out)]
-            )
-        tt_out = tt_out_tmp
-    elif llm_mode == "decode":
-        tt_out = [tt_out_torch.squeeze(1).transpose(0, 1) for tt_out_torch in tt_tensors_to_torch_tensors(tt_out)]
-        tt_out = torch.concat(tt_out)
-
-    # check outputs ----------------------------------------------------------------------
-    does_pass = True
-    tt_out_tmp = tt_out.type(pytorch_out.dtype)
-    _, _, device_pcc, pcc_str = get_atol_rtol_pcc(pytorch_out, tt_out_tmp)
-    logger.info(f"Output: {pcc_str}")
-    if device_pcc < expected_pccs[0]:
-        does_pass = False
-        logger.warning(f"Output PCC {device_pcc} is lower than {expected_pccs[0]}")
-    if device_pcc > (expected_pccs[0] + 0.01):
-        does_pass = False
-        logger.warning(f"Output PCC {device_pcc} is higher than {expected_pccs[0]}. Please update the expected PCC")
-
-    reference_logits = pytorch_out.view(global_batch * seq_len, -1).float().detach().numpy()
-    eval_logits = tt_out.view(global_batch * seq_len, -1).float().detach().numpy()
-    reference_top1 = np.argmax(reference_logits, axis=-1)
-    top1_acc = top_k_accuracy_score(reference_top1, eval_logits, k=1, labels=np.arange(eval_logits.shape[-1]))
-    top5_acc = top_k_accuracy_score(reference_top1, eval_logits, k=5, labels=np.arange(eval_logits.shape[-1]))
-    logger.info(f"Top-1 Accuracy: {top1_acc}")
-    logger.info(f"Top-5 Accuracy: {top5_acc}")
-
-    device_pcc_k = 1.0
-    device_pcc_v = 1.0
-    for i in range(num_layers):
-        if llm_mode == "prefill":
-            pytorch_layer_pres = (pytorch_layer_present[i][0].squeeze(1), pytorch_layer_present[i][1].squeeze(1))
-            tt_layer_pres = concat_device_out_layer_present(num_devices, tt_layer_present[i], kv_len)
-        elif llm_mode == "decode":
-            pytorch_layer_pres = (
-                pytorch_layer_present[i][0].squeeze(1)[:, kv_cache_len, :],
-                pytorch_layer_present[i][1].squeeze(1)[:, kv_cache_len, :],
-            )
-            tt_layer_pres = concat_device_out_layer_present(
-                num_devices, tt_layer_present[i], kv_cache_len, end_idx_only=True
-            )
-        tt_layer_pres_0 = tt_layer_pres[0].type(pytorch_layer_pres[0].dtype)
-        _, _, device_pcc, pcc_str = get_atol_rtol_pcc(pytorch_layer_pres[0], tt_layer_pres_0)
-        logger.info(f"K Cache Layer {i}: {pcc_str}")
-        device_pcc_k = min(device_pcc_k, device_pcc)
-
-        tt_layer_pres_1 = tt_layer_pres[1].type(pytorch_layer_pres[1].dtype)
-        _, _, device_pcc, pcc_str = get_atol_rtol_pcc(pytorch_layer_pres[1], tt_layer_pres_1)
-        logger.info(f"V Cache Layer {i}: {pcc_str}")
-        device_pcc_v = min(device_pcc_v, device_pcc)
-
-    logger.info(f"Device PCC K: {device_pcc_k}")
-    logger.info(f"Device PCC V: {device_pcc_v}")
-
-    if device_pcc_k < expected_pccs[1]:
-        does_pass = False
-        logger.warning(f"K Cache PCC {device_pcc_k} is lower than {expected_pccs[1]}")
-    if device_pcc_k > (expected_pccs[1] + 0.01):
-        does_pass = False
-        logger.warning(f"K Cache PCC {device_pcc_k} is higher than {expected_pccs[1]}. Please update the expected PCC")
-
-    if device_pcc_v < expected_pccs[2]:
-        does_pass = False
-        logger.warning(f"V Cache PCC {device_pcc_v} is lower than {expected_pccs[2]}")
-    if device_pcc_v > (expected_pccs[2] + 0.01):
-        does_pass = False
-        logger.warning(f"V Cache PCC {device_pcc_v} is higher than {expected_pccs[2]}. Please update the expected PCC")
-
-    profiler.print()
-
-    comment = f"num_devices={num_devices}_kv_cache_len={kv_cache_len}_seq_len={seq_len}_num_layers={num_layers}_config={model_config_str}_async={async_mode}"
-    cpu_time = profiler.get("hugging_face_reference_model")
-    first_iter_time = profiler.get("first_model_run_with_compile")
-    second_iter_time = profiler.get("model_run_for_inference")
-    expected_compile_time = 44
-    prep_perf_report(
-        model_name=f"Falcon_{llm_mode}_{comment}",
-        batch_size=batch,
-        inference_and_compile_time=first_iter_time,
-        inference_time=second_iter_time,
-        expected_compile_time=expected_compile_time,
-        expected_inference_time=expected_inference_time,
-        comments=comment,
-        inference_time_cpu=cpu_time,
-    )
-
-    compile_time = first_iter_time - second_iter_time
-    logger.info(f"falcon {comment} inference time: {second_iter_time}")
-    logger.info(f"falcon {comment} compile time: {compile_time}")
-
-    if does_pass:
-        logger.info("Falcon PCC Check Passed!")
-    else:
-        logger.warning("Falcon PCC Check Failed!")
-        assert (
-            does_pass
-        ), f"Output PCC, k_cache_pcc, or v_cache_pcc is either lower or higher than {expected_pccs}. See earlier warnings for more details."
 
 
 @pytest.mark.parametrize(
@@ -451,6 +101,7 @@ class TestParametrized:
             tt_cache_path,
             model_location_generator,
             expected_inference_time,
+            e2e_perf=True,
         )
 
     def run_perf_wh_bare_metal(
@@ -502,6 +153,7 @@ class TestParametrized:
             model_location_generator,
             expected_inference_time,
             async_mode,
+            e2e_perf=True,
         )
 
     @pytest.mark.models_performance_bare_metal
