+++ models/demos/falcon7b/tt/falcon_model.py
@@ -8,6 +8,8 @@ from typing import List, Optional, Tuple
 import torch
 import ttnn
 
+import tt_lib
+
 from models.demos.falcon7b.tt.falcon_decoder import TtFalconDecoderLayer
 from models.demos.falcon7b.tt.model_utils import get_weights_cached
 from models.utility_functions import nearest_32, torch_tensors_to_tt_tensors
@@ -250,6 +252,7 @@ class TtFalconModelShared(torch.nn.Module):
         layer_past: Optional[Tuple[Tuple[ttnn.experimental.tensor.Tensor]]] = None,
         layer_past_len: int = 0,
         use_cache: bool = False,
+        device_perf_run: bool = False,
     ) -> ttnn.experimental.tensor.Tensor:
         # Convert input tokens to embeddings
         input_embeddings = [
@@ -280,6 +283,10 @@ class TtFalconModelShared(torch.nn.Module):
             )
             presents += (layer_output[1],)
             layer_output = layer_output[0]
+
+            if device_perf_run and idx % 8 == 0:
+                for i in range(self.num_devices):
+                    tt_lib.device.DumpDeviceProfiler(layer_output[i].device())
 
         # apply final norm layer
         for i in range(self.num_devices):
