--- tests/nightly/wh_b0_only_eth/models/demos/falcon7b/tests/ci/test_falcon_end_to_end_prefill.py
@@ -1,1 +1,0 @@
-../../../../../../../../models/demos/falcon7b/tests/ci/test_falcon_end_to_end_prefill.py
